task-06-graphql/python-server/requirements.txt
cached-property==1.5.2
click==7.1.2
graphql-core==3.1.3
h11==0.12.0
hupper==1.10.2
Pygments==2.8.0
python-dateutil==2.8.1
python-multipart==0.0.5
six==1.15.0
starlette==0.14.2
strawberry-graphql==0.48.0
typing-extensions==3.7.4.3
uvicorn==0.13.4
